id should be long, not int

+++ src/main/java/models/User.java
@@ -28,7 +28,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     private String firstname;
     private String lastname;
